added javadocs and separated sessions
